refactor: split Settings page into focused panel components (Phase 4)
Settings was a 1016-line god page (drifted up from 904 in the brief)
mixing 30+ useState calls, 4 useEffect hooks, ~10 panels, and 8
async handlers. This PR keeps the page as a thin orchestrator
(99 lines) and moves the 9 panels into focused, self-contained
components. Behavior is unchanged — the existing 103 vitest tests
all pass.

Settings decomposes more cleanly than VolunteerDashboard (PR #125)
because the panels are independent forms — most state is panel-local
and each panel can manage its own form state and only needs `profile`
+ `onSaved` from the page. The "panels own their state" pattern means
the orchestrator is genuinely thin.

New files:
  src/lib/password-strength.ts                   17  Pure helper
  src/components/settings/types.ts               30  SettingsProfile
                                                     extension type
  src/components/settings/AvatarUploadField.tsx 119  Avatar upload + remove
  src/components/settings/ProfilePanel.tsx      162  Name/email/phone/
                                                     emergency/DOB form
  src/components/settings/ParentalConsentPanel  122  Minor consent (owns
                                                     its consent fetch)
  src/components/settings/UsernamePanel.tsx     133  Debounced availability
                                                     + save
  src/components/settings/PasswordPanel.tsx      73  Change password
  src/components/settings/MfaPanel.tsx          223  Enroll/verify/remove/
                                                     regenerate; renders
                                                     BackupCodesDialog
  src/components/settings/BackupCodesDialog      98  One-time codes display
  src/components/settings/NotificationsPanel    153  3 channel + 5 per-type
                                                     switches
  src/components/settings/DeleteAccountPanel    105  Danger zone (non-admin)

Cross-panel state lifted to the page (preserves pre-refactor UX):
  - phone, emergencyPhone — read by NotificationsPanel to gate the
    SMS toggle on un-saved phone changes
  - isMinor — drives ParentalConsentPanel visibility on un-saved DOB
    changes

Page: 1016 → 99 (-90%)
Total: 1016 → 1334 (+318 net; ~900 lines of logic moved out).

Type strategy:
- SettingsProfile extension type covers 8 unmodelled columns
  (date_of_birth, is_minor, username, notif_shift_reminders,
  notif_new_messages, notif_milestone, notif_document_expiry,
  notif_booking_changes). Page bridges with `profile as unknown as
  SettingsProfile` once at the panel boundary.
- Existing (supabase as any).rpc() calls retained verbatim (matches
  documented eslint.config.js pattern).
- 4 new boundary casts introduced in new code, all at supabase
  type-system boundaries:
  * 2× `as never` on profiles.update payload (Json type vs literal
    object types — same pattern as booking-actions.ts in PR #125)
  * 1× `data[0] as any` in ParentalConsentPanel for embedded-select
    response (PostgREST shape doesn't survive generated types)
  * 1× `profile as unknown as SettingsProfile` in the page bridge

Follow-up issue filed: #126 (MFA backup code count stale after code
consumption — pre-existing bug surfaced but not fixed in this
behavior-preservation refactor).

Verification: tsc --build clean, eslint --max-warnings=0 clean,
103/103 vitest passes.

Second of five Phase 4 god-component refactors. ManageShifts.tsx
(~700 lines) is next.
